requirements.txt
qdrant-client
sentence-transformers
numpy
pillow
pytesseract